Bring laps text running

+++ RaceOverlay/Data/Mapper.cs
@@ -12,8 +12,8 @@ public class Mapper
         iRacingData data = new();
         
         // Map Session Data
-        data.SessionData.TimeLeft = irsdkSharper.Data.GetFloat("SessionTimeRemain");
-        data.SessionData.TimeTotal = irsdkSharper.Data.GetFloat("SessionTimeTotal");
+        data.SessionData.TimeLeft = irsdkSharper.Data.GetDouble("SessionTimeRemain");
+        data.SessionData.TimeTotal = irsdkSharper.Data.GetDouble("SessionTimeTotal");
         data.SessionData.LapsLeft = irsdkSharper.Data.GetInt("SessionLapsRemain");
         data.SessionData.LapsTotal = irsdkSharper.Data.GetInt("SessionLapsTotal");
         data.SessionData.LapsLeftEstimated = irsdkSharper.Data.GetInt("SessionLapsRemainEx");
